Add test for initialization retry testing
Add the extra CameraX initialization test which didn't use any
CameraUtil in the testing package to check the camera framework
before testing.
The goal of this test is to initialize the CameraX even when camera
HAL is in a bad state to ensure the retry mechanism works properly.

Add the extra 2 devices to the list in IncompleteCameraListQuirk

Test: manually run CameraXInitTest on lab devices
Change-Id: Iba44e90b6323a86e9d93525ff8b74d4dfba6d4d0