Remember the graph that is passed to a NavHost
Currently if there is a compose caused by a change in state outside of
the NavHost, the NavHost resets the state of the graph. This is because
the recompose will create a new instance of the NavGraph (or
NavGraphBuilder) and that new instance will be detected as a new graph
and be set on the NavController.

To avoid this, we should remember the initial NavGraph, and maintain the
same graph for the life of the NavHost.

Test: NavHostTest
Bug: 172185423
Change-Id: I0190753ead4b79ffe4c58591e8721307ca4c8f28