In OutputFileOptions, create new Metadata as default value.
Previously, the default EMPTY_METADATA is essentially a singleton. If the user creates a OutputFileOptions without Metadata, then proceed to edit the Metadata, it will change the default value.

This might not be an issue for most of the core users. But it happens in CameraController where we override the horizontal flag.

Bug: 171795734
Test: manual test and ./gradlew bOS
Change-Id: Ib3a88fc396d4f1cdc5bb3cc43a92086a852c4ae9