Add default method for execPerConnectionSQL() in SupportSQLiteDatabase
SQLiteDatabase#execPerConnectionSQL() was added in API 30 and cannot be backported but a new default method for it can be declared so those who provide the androidx.sqlite interfaces with unbundled SQLite bindings can implement it. For the default framework implementation of the interface the method will throw an UnsupportedOperationException if the API level is lower than 30. To check if the operation is supported users should call isExecPerConnectionSQLSupported(). This allow us to not add @RequireApi(30) which would prevent unbundled implementations to not offer the functionality to lower API levels.

Bug: 172270145
Test: N/A
Relnote: Add default method for execPerConnectionSQL() in SupportSQLiteDatabase
Change-Id: I86326d81aad82c1efccbb2087c439655c9b4100f