Skip loading non-final versions of WM Extensions
To make it clear that a non-final imlplementations of
androidx.window.extensions must not be included on device builds,
this CL skips the Extension loading on release builds of
androidx.window even before checking the API match.

Bug: 172620880
Test: Build a release version of WM Jetpack, check if
androidx.window.extensions alpha is not picked up.
Test: Build a debug version of WM Jetpack, check if
androidx.window.extensions alpha is picked up (for testing).

Change-Id: I6c6ab1beff91e435fc414d43e71f85839fe45954