Test failures: destroy fragment after test
Not sure how exactly it works, but my guess is on lower end devices, the camera needs longer time to open after a few on/offs. So shutting down the lifecycle proactively helps. It also works by increasing timeout duration but it's less elegant.

This should fix a greate number of the current failures in view artifact.

Bug: 171956775, 171794776
Test: manual test and ./gradlew bOS
Change-Id: I9e833646b4c88895f151228b02fbc9cd7cd0eff7