CameGraph submit3A implementation.
- This uses the RequestProcessor's submit method that takes a map of
parameters and build the request bases on the latest repeating request.
- Extracted some common methods between update3A and submit3A
- If RequestProcessor encounters any error then this method returns
and invalid frame number i.e. -1 and FAILED status code.
- Added a method in Listener3A to remove a listener. This is useful
in case when RequestProcessor's submit method returns false and we
don't want to keep a stale listener forever attached to the global pool
of 3a listeners.

Test: ./gradlew :camera:camera-camera2-pipe:test
      ./gradlew :camera:camera-camera2-pipe:ktlint

Tweaked the sample app to invoke the submit3A method on shutter
press and added the log to confirm the frame number at which it was
applies

Change-Id: I492a56d02f5f8cff431c6fa36f07f56af43efb68