Make the Serializer's OutputStream.close() a no-op.
Instead of throwing IllegalStateException when the OutputStream is closed, let's make it a no-op. We need the underlying FileOutputStream to be open after the Serializer has completedso we can get access to and sync the file descriptor. Previously we required that the OutputStream was not closed, but this can be problematic, for example: b/173037611

Bug: 173037611
Relnote: We now allow (but don't require) the OutputStream passed to Serializer.writeTo() to be closed.
Test: Updated the original output stream closing test.
Change-Id: I5c9bffd60bd12179a496e4589b010e31d00d8a5a