Ensure canceled Animations call complete()
Exiting fragments that have their Animation canceled, get an onCancel
callback from their CancelationSignal in DefaultSpecialEffects, but the
call to clearAnimation(), never actually calls onAnimationEnd, so the
Operation never completes, which means views cannot be properly cleaned
up cause the Fragment is never destroyed.

Instead of relying on the view system to complete the SpecialEffect, we
should just handle the cancelation in onCancel directly. This way we can
ensure that the fragments are cleaned appropriately.

RelNote: "Fixed an issue with the new state manager that prevented fragments with an exiting `Animation` from being properly destroyed."

Test: tested in sample app
Fixes: 171745757
Change-Id: I83d6584916bb3425f552d73a17e1ab938f00190d