Cancel animators by using animator.end()
When handling animators, DefaultSpecialEffectsController currently uses
animator.cancel() to cancel Animators. The animator.cancel calls leaves
the target view in whatever state it is in during the cancel call and
for entering fragments this means they are left with an alpha of 0,
which means they do not show up.

The original reason for using animator.cancel() was that in the case
where a running animator was interrupted, it could begin the new
animator from whatever state the view was in at the moment it was
canceled, but in practice it start state of the new animator takes over
and it looks as if the animator too the view to the end state anyway.

Switching to use animator.end() when canceling fixes the above problem
by changing the state of the view to the final state which means the
alpha is properly set to 1.0 for entering fragments and this
behavior also aligns Animators with Animations.

RelNote: "Fixed an issue with the new state manager where the view would not reach its final state if you interrupt a fragment transition that was using an `Animator` or one of the `TRANSIT_FRAGMENT_` options."
Test: FragmentAnimatorTest
Bug: 169874632

Change-Id: I92426d8572f9ca71c053945ecc2afd839efbc058