Fix improper animations after pop
When combining pop and replace operations in the same fragment
transaction and using the new state manager, the fragment in the
transaction being popped does not have a view and therefore its
animation is ignored.

We should ensure we consider any animations that are available, even if
the fragment has no view.

RelNote: "Combining pop and replace operations in the same fragment
transaction will now show the proper animations"
Test: removingFragmentAnimationChange
Bug: 170328691

Change-Id: Ifd4e49cd9874a9ac0673083c27ef3c9765bbfa04