Do not invalidate if state has not changed.
The watch face is notified every time the battery status changes. However, in
most cases, this does not affect the watch face state, as we are only store
whether the battery is low and not charging. If a broadcast is delivered but
the state has not changed, we can simply ignore the update and not schedule a
redraw.

Bug: 175707598
Test: manually sending broadcast + ./gradlew :wear:wear-watchface:test
Change-Id: Icbcb8e261099004ba7ee69e061bc463e9f355496