Register for the correct broadcasts.
The code was registering for ACTION_BATTERY_STATUS but then acting on
the ACTION_BATTERY_LOW, ACTION_BATTERY_OKAY, and
ACTION_POWER_CONNECTED actions, which it did not register for, and
hence were never delivered.

This commit changes the code to register and notify for these three
broadcasts individually, which also removes the case where the
broadcast does not match any of the actions and it is ignored.

Bug: 177041795
Test: manually sending broadcast + ./gradlew :wear:wear-watchface:test
Change-Id: Ia5515e22e0f9ea4da6ee8c2c39df1df28f7a2baa