Use immutable collection for KSP
In KSP, when a java type is resolved, it is resolved to its mutable
counterpart if available (e.g. java.util.List resolves to
kotlin.collections.MutableList). Unfortunately, this doesn't work well
when we want to just check for assignability for known types.
Specifically in this case, we want to know if a query parameter is a
collection or not and we don't care for mutability.

This CL updates Context to specify immutable collection qName if the
processor is KSP. We might want to map kotlin types to their immutable
variations but I'm not sure about that for now and that is also not what
kotlin does by default (well, immutable ones do not exist in javac
anyways so it is undefined territory).

To test, I've added a new test instead of updating query method
parameter test to run w/ KSP as it requires a major refactoring and
having a base KSP processor API.

Bug: 160322705
Test: TypeAdapterStoreTest.findQueryParameterAdapter_collections

Change-Id: I6de68d474dc24b38c662fed753bf6eb9af2ba101