Replace channelFlow with a basic non-experimental implementation
This CL replaces usages of ChannelFlow with a custom implementation that
does not depend on experimental coroutines APIs.

It is not completely the same implementation as we lose the benefits of
operator fusing. To get close, I'm internally using a channel(0)
followed by a buffer operator to use DEFAULT buffer. This is close to
what happens with channelFlow except the intermediate channel causes 1
more value to be produced even when buffer(0) is used on the resulted
stream.

I also had a make 2 changes on PageFetcherSnapshotTest. Both of these
cases were asserting ordering in a place where there is no order
guarantees.

Relnote: N/A
Bug: 171708323
Test: SimpleChannelFlowTest, existing tests
Change-Id: I474e1eed48072739a3f9d84e2c592e3f1e491cfe