Fix PreviewTransformationDeviceTest#correctTextureViewWith90Rotation() test failures on some devices.
On some devices, calling the same Matrix#mapPoints() function may have different results. It might be caused by the deviation of float number calculation and make the result checking failed. Because the Matrix#mapPoints() here is used for the view points transformation and view points are actually integers, changing to do the result checking with IntArray in the unit test to avoid such kind of test failures.

Bug: 176884913
Test: PreviewTransformationDeviceTest
Change-Id: I6cd26819e4cefa80cd2c259c2a2f875b899cda68