Properly check permissions for globalQuery.
Take the UID and the package name of the caller during globalQuery. The
UID is used to verify if the caller has access to platform surfaceable
data. The package name is used to allow access to the caller's own
data. For 3p access, the UID must be associated with a specified
PackageIdentifier.

Bug: 169883602
Test: ./gradlew appsearch:appsearch:connectedCheck --info --daemon
  ./gradlew appsearch:appsearch-local-storage:connectedCheck --info
  --daemon
Change-Id: I2eebc69f4f5ba12b7418df519474fe3c1ad74b22