Fix visual artifacts when using hide
When doing a hide operation, because the fragment view is not destroyed
at the end, we cannot put an animation in an AnimationSet or we will
get a visual artifact.

We need to treat hide operations with animation the same way we treat
adding fragment with animations. We start the animation and then
immediately complete the special effect.

RelNote: "Fixed issue where doing a hide operation with an Animation
would cause the hiding fragment to flash at the end of the animation."
Test: tested in sample app
Bug: 175417675

Change-Id: I57e2282d6c16dcbb403379d67c5e747379db58b1